+++ dwara-core/src/main/java/org/ishafoundation/dwaraapi/storage/archiveformat/tar/response/TarResponseParser.java
@@ -68,7 +68,7 @@ public class TarResponseParser {
 				
 				file.setFilePathName(filePathName);
 				file.setFileSize(Long.parseLong(fileSizeAsString));
-				file.setArchiveBlock(Integer.parseInt(archiveBlockOffset));
+				file.setArchiveBlock(Long.parseLong(archiveBlockOffset));
 
 				tarResponse.getFileList().add(file);
 			}
